Move DisplayIdentifier into its own file
