+++ api/src/main/scala/uk/ac/wellcome/platform/api/models/DisplayWork.scala
@@ -89,22 +89,3 @@ case object DisplayWork {
     )
   }
 }
-
-@ApiModel(
-  value = "Identifier",
-  description =
-    "A unique system-generated identifier that governs interaction between systems and is regarded as canonical within the Wellcome data ecosystem."
-)
-case class DisplayIdentifier(
-  @ApiModelProperty(value =
-    "Relates a Identifier to a particular authoritative source identifier scheme: for example, if the identifier is MS.49 this property might indicate that this identifier has its origins in the Wellcome Library's CALM archive management system.") identifierScheme: String,
-  @ApiModelProperty(value = "The value of the thing. e.g. an identifier") value: String) {
-  @ApiModelProperty(readOnly = true, value = "A type of thing")
-  @JsonProperty("type") val ontologyType: String = "Identifier"
-}
-
-object DisplayIdentifier {
-  def apply(sourceIdentifier: SourceIdentifier): DisplayIdentifier =
-    DisplayIdentifier(identifierScheme = sourceIdentifier.identifierScheme,
-                      value = sourceIdentifier.value)
-}
